+++ rewards/protobuf/rewards.proto
@@ -8,45 +8,6 @@ service Rewards {
     rpc GetParameters(ParametersRequest) returns(ParametersResponse) {}
 }
 
-enum Currency {
-	AED = 0;
-	ARS = 1;
-	AUD = 2;
-	BAT = 3;
-	BCH = 4;
-	BRL = 5;
-	BTC = 6;
-	BTG = 7;
-	CAD = 8;
-	CHF = 9;
-	CNY = 10;
-	DASH = 11;
-	DKK = 12;
-	ETH = 13;
-	EUR = 14;
-	GBP = 15;
-	HKD = 16;
-	ILS = 17;
-	INR = 18;
-	JPY = 19;
-	KES = 20;
-	LBA = 21;
-	LTC = 22;
-	MXN = 23;
-	NOK = 24;
-	NZD = 25;
-	PHP = 26;
-	PLN = 27;
-	SEK = 28;
-	SGD = 29;
-	USD = 30;
-	XAG = 31;
-	XAU = 32;
-	XPD = 33;
-	XPT = 34;
-	XRP = 35;
-}
-
 // ParametersRequest - Primary type of the parameters
 // get endpoint request.
 message ParametersRequest {
@@ -73,3 +34,213 @@ message Tips {
     repeated double defaultTipChoices = 1;
     repeated double defaultMonthlyChoices = 2;
 }
+
+enum Currency {
+    AED = 0;
+    AFN = 1;
+    ALL = 2;
+    AMD = 3;
+    ANG = 4;
+    AOA = 5;
+    ARS = 6;
+    AUD = 7;
+    AWG = 8;
+    AZN = 9;
+    BAM = 10;
+    BBD = 11;
+    BDT = 12;
+    BGN = 13;
+    BHD = 14;
+    BIF = 15;
+    BMD = 16;
+    BND = 17;
+    BOB = 18;
+    BRL = 19;
+    BSD = 20;
+    BTC = 21;
+    BTN = 22;
+    BWP = 23;
+    BYN = 24;
+    BZD = 25;
+    CAD = 26;
+    CDF = 27;
+    CHF = 28;
+    CLF = 29;
+    CLP = 30;
+    CNH = 31;
+    CNY = 32;
+    COP = 33;
+    CRC = 34;
+    CUC = 35;
+    CUP = 36;
+    CVE = 37;
+    CZK = 38;
+    DJF = 39;
+    DKK = 40;
+    DOP = 41;
+    DZD = 42;
+    EGP = 43;
+    ERN = 44;
+    ETB = 45;
+    EUR = 46;
+    FJD = 47;
+    FKP = 48;
+    GBP = 49;
+    GEL = 50;
+    GGP = 51;
+    GHS = 52;
+    GIP = 53;
+    GMD = 54;
+    GNF = 55;
+    GTQ = 56;
+    GYD = 57;
+    HKD = 58;
+    HNL = 59;
+    HRK = 60;
+    HTG = 61;
+    HUF = 62;
+    IDR = 63;
+    ILS = 64;
+    IMP = 65;
+    INR = 66;
+    IQD = 67;
+    IRR = 68;
+    ISK = 69;
+    JEP = 70;
+    JMD = 71;
+    JOD = 72;
+    JPY = 73;
+    KES = 74;
+    KGS = 75;
+    KHR = 76;
+    KMF = 77;
+    KPW = 78;
+    KRW = 79;
+    KWD = 80;
+    KYD = 81;
+    KZT = 82;
+    LAK = 83;
+    LBP = 84;
+    LKR = 85;
+    LRD = 86;
+    LSL = 87;
+    LYD = 88;
+    MAD = 89;
+    MDL = 90;
+    MGA = 91;
+    MKD = 92;
+    MMK = 93;
+    MNT = 94;
+    MOP = 95;
+    MRO = 96;
+    MRU = 97;
+    MUR = 98;
+    MVR = 99;
+    MWK = 100;
+    MXN = 101;
+    MYR = 102;
+    MZN = 103;
+    NAD = 104;
+    NGN = 105;
+    NIO = 106;
+    NOK = 107;
+    NPR = 108;
+    NZD = 109;
+    OMR = 110;
+    PAB = 111;
+    PEN = 112;
+    PGK = 113;
+    PHP = 114;
+    PKR = 115;
+    PLN = 116;
+    PYG = 117;
+    QAR = 118;
+    RON = 119;
+    RSD = 120;
+    RUB = 121;
+    RWF = 122;
+    SAR = 123;
+    SBD = 124;
+    SCR = 125;
+    SDG = 126;
+    SEK = 127;
+    SGD = 128;
+    SHP = 129;
+    SLL = 130;
+    SOS = 131;
+    SRD = 132;
+    SSP = 133;
+    STD = 134;
+    STN = 135;
+    SVC = 136;
+    SYP = 137;
+    SZL = 138;
+    THB = 139;
+    TJS = 140;
+    TMT = 141;
+    TND = 142;
+    TOP = 143;
+    TRY = 144;
+    TTD = 145;
+    TWD = 146;
+    TZS = 147;
+    UAH = 148;
+    UGX = 149;
+    USD = 150;
+    UYU = 151;
+    UZS = 152;
+    VEF = 153;
+    VES = 154;
+    VND = 155;
+    VUV = 156;
+    WST = 157;
+    XAF = 158;
+    XAG = 159;
+    XAU = 160;
+    XCD = 161;
+    XDR = 162;
+    XOF = 163;
+    XPD = 164;
+    XPF = 165;
+    XPT = 166;
+    YER = 167;
+    ZAR = 168;
+    ZMW = 169;
+    ZWL = 170;
+    ADA = 171;
+    ATOM = 172;
+    BAT = 173;
+    BCH = 174;
+    BTG = 175;
+    -DAI = 176;
+    DASH = 177;
+    DCR = 178;
+    DGB = 179;
+    DOGE = 180;
+    EOS = 181;
+    ETH = 182;
+    IOTA = 183;
+    LBA = 184;
+    LINK = 185;
+    LTC = 186;
+    NANO = 187;
+    NEO = 188;
+    OMG = 189;
+    STORM = 190;
+    TRX = 191;
+    -TUSD = 192;
+    UPBTC = 193;
+    UPEUR = 194;
+    UPT = 195;
+    UPUSD = 196;
+    -USDC = 197;
+    -USDT = 198;
+    VOX = 199;
+    XEM = 200;
+    XLM = 201;
+    XRP = 202;
+    ZIL = 203;
+    ZRX = 204;
+    BCC = 205;
+}
+
